+++ dcicutils/misc_utils.py
@@ -997,6 +997,96 @@ def to_integer(value: str, fallback: Optional[Any] = None, strict: bool = False)
     return fallback
 
 
+_MULTIPLIER_K = 1000
+_MULTIPLIER_M = 1000 * _MULTIPLIER_K
+_MULTIPLIER_G = 1000 * _MULTIPLIER_M
+_MULTIPLIER_T = 1000 * _MULTIPLIER_G
+
+_MULTIPLIER_SUFFIXES = {
+    "K": _MULTIPLIER_K,
+    "Kb": _MULTIPLIER_K,
+    "KB": _MULTIPLIER_K,
+    "M": _MULTIPLIER_M,
+    "Mb": _MULTIPLIER_M,
+    "MB": _MULTIPLIER_M,
+    "G": _MULTIPLIER_G,
+    "Gb": _MULTIPLIER_G,
+    "GB": _MULTIPLIER_G,
+    "T": _MULTIPLIER_T,
+    "Tb": _MULTIPLIER_T,
+    "TB": _MULTIPLIER_T
+}
+
+def to_number(value: str,
+              allow_prefix: bool = True,
+              allow_commas: bool = False,
+              allow_suffix: bool = False,
+              allow_float: bool = False) -> Optional[Union[int, float]]:
+
+    """
+    Converts the give string value to an int, or possibly float if allow_float is True.
+    If allow_prefix is True (default: True) then allow/respect a preceeding '+' or '-'.
+    If allow_commas is True (default: False) then allow commas (only) every three digits.
+    If allow_suffix is True (default: False) allow any of K, Kb, KB; or M, Mb, MB; or
+    G, Gb, or GB; or T, Tb, TB, to mean multiply value by one thousand; one million;
+    one billion; or one trillion; respectively. If allow_float is True (default: False)
+    allow the value to be floating point (i.e. with a decimal point and a fractional part),
+    in which case the returned value will be of type float rather than int.
+    If the string is not well formated then returns None.
+    """
+    if not (isinstance(value, str) and (value := value.strip())):
+        return value if isinstance(value, (int, float)) else None
+
+    value_multiplier = 1
+    value_negative = False
+    value_fraction = None
+
+    if allow_prefix is True:
+        if value.startswith("-"):
+            if not (value := value[1:].strip()):
+                return None
+            value_negative = True
+        elif value.startswith("+"):
+            if not (value := value[1:].strip()):
+                return None
+
+    if allow_suffix is True:
+        for suffix in _MULTIPLIER_SUFFIXES:
+            if value.endswith(suffix):
+                value_multiplier *= _MULTIPLIER_SUFFIXES[suffix]
+                if not (value := value[:-len(suffix)].strip()):
+                    return None
+                break
+
+    if allow_float is True:
+        if dot_index := value.rindex("."):
+            if value_fraction := value[dot_index + 1:].strip():
+                try:
+                    value_fraction = float(f"0.{value_fraction}")
+                except Exception:
+                    return None
+            value = value[:dot_index].strip()
+        pass
+
+    if allow_commas is True:
+        if not re.fullmatch(r"(-?\d{1,3}(,\d{3})*)", value):
+            return None
+        value = value.replace(",", "")
+
+    if not value.isdigit():
+        return None
+
+    result = int(value) * value_multiplier
+
+    if value_fraction:
+        result += value_fraction
+
+    if value_negative:
+        result = -result
+
+    return result
+
+
 def to_float(value: str, fallback: Optional[Any] = None) -> Optional[Any]:
     try:
         return float(value)